% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{adjacent}
\alias{adjacent}
\title{Adjacent vertices.}
\usage{
adjacent(st, vertices = NULL)
}
\arguments{
\item{st}{a simplex tree.}

\item{vertices}{vertex ids.}
}
\description{
Returns a list of vertex ids that are immediately adjacent to a given vertex. If a given vertex
does not have any adjacencies, a vector of length 0 is returned.
}
\examples{
st <- simplex_tree(1:3)
st \%>\% adjacent(2) 
# 1 3
}
