% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{as.list.st_traversal}
\alias{as.list.st_traversal}
\title{as.list.st_traversal}
\usage{
\method{as.list}{st_traversal}(x, ...)
}
\arguments{
\item{x}{traversal object.}

\item{...}{unused.}
}
\description{
as.list.st_traversal
}
