% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{as_XPtr}
\alias{as_XPtr}
\title{Convert to external pointer}
\description{
Exports the simplex tree as an \code{externalptr} (i.e. \code{Rcpp::Xptr}) for passing to and from C++ and R. 
This method does not register a finalizer. An example is given below using the Rcpp \emph{depends} attribute.
}
\examples{

## Below is an example of casting an XPtr created in R to a SimplexTree type in C++. 
\dontrun{
// my_source.cpp
#include "Rcpp.h"
// [[Rcpp::depends(simplextree)]]
#include "simplextree.h"

[[Rcpp::export]]
void print_tree(SEXP stree){
 Rcpp::XPtr<SimplexTree> stree_ptr(stree);
 stree_ptr->print_tree();
}
}
## Pass to Rcpp as follows
st <- simplextree::simplex_tree()
print(st$as_XPtr())
}
