% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cech.R
\name{cech}
\alias{cech}
\title{Cech complex}
\usage{
cech(x, radius = "default", dim = 1L, dx = NULL)
}
\description{
Computes the Cech complex up to some radius and dimension.
}
\details{
The Cech complex at \code{radius} is defined as the abstract simplicial complex which contains a k-simplex
for every (k+1) non-empty intersection of balls with size \code{radius}. The Cech complex is equivalently defined
as the nerve complex of the cover formed by placing \code{radius}-balls at every point in \code{x}.
}
