% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{clear}
\alias{clear}
\title{Clears the simplex tree}
\description{
Removes all simplices from the simplex tree, except the root node.
}
\examples{
st <- simplex_tree()
st$insert_simplex(1:3)
st$clear()
print(st) ## < empty simplex tree >
}
