% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{clone}
\alias{clone}
\title{Clone a simplex tree}
\usage{
clone(st)
}
\arguments{
\item{st}{a simplex tree.}
}
\description{
Performs a deep-copy on the supplied simplicial complex.
}
\details{
A clone is produced by serializing the input simplex tree \code{st}
  and deserializing it into a new simplex tree that is then returned.
}
\seealso{
Other serialization methods: 
\code{\link{deserialize}()},
\code{\link{serialize}()}
}
\concept{serialization methods}
