% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{coface_roots}
\alias{coface_roots}
\title{Coface roots traversal}
\usage{
coface_roots(st, sigma)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{sigma}{simplex to start the traversal at.}
}
\value{
a traversal (object of class \code{"st_traversal"}).
}
\description{
Generates a coface roots traversal on the simplex tree.

The coface roots of a given simplex \code{sigma} are the roots
of subtrees in the trie whose descendents (including the roots themselves)
are cofaces of \code{sigma}. This traversal is more useful when used in
conjunction with other traversals, e.g. a \emph{preorder} or
\emph{level order} traversal at the roots enumerates the cofaces of
\code{sigma}.
}
\seealso{
Other traversals: 
\code{\link{cofaces}()},
\code{\link{faces}()},
\code{\link{k_simplices}()},
\code{\link{k_skeleton}()},
\code{\link{level_order}()},
\code{\link{link}()},
\code{\link{maximal}()},
\code{\link{preorder}()},
\code{\link{traverse}()}
}
\concept{traversals}
