% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{cofaces}
\alias{cofaces}
\title{Coface traversal}
\usage{
cofaces(st, sigma)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{sigma}{simplex to start the traversal at.}
}
\value{
a traversal (object of class \code{"st_traversal"}).
}
\description{
Generates a coface traversal on the simplex tree.
}
\seealso{
Other traversals: 
\code{\link{coface_roots}()},
\code{\link{faces}()},
\code{\link{k_simplices}()},
\code{\link{k_skeleton}()},
\code{\link{level_order}()},
\code{\link{link}()},
\code{\link{maximal}()},
\code{\link{preorder}()},
\code{\link{traverse}()}
}
\concept{traversals}
