% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{collapse}
\alias{collapse}
\title{Elementary collapse}
\usage{
$collapse(tau, sigma)
}
\arguments{
\item{tau}{a k-length vector of vertex ids representing a (k-1)-simplex.}

\item{sigma}{a l-length vector of vertex ids representing a (l-1)-simplex.}
}
\value{
boolean indicating whether \code{tau} is a coface of \code{sigma}, and the two were removed.
}
\description{
Performs an elementary collapse.
}
\details{
This function performs an \emph{elementary collapse} in the sense described by (1), which is 
summarized here. A simplex \eqn{\sigma} is said to be collapsible through one of its faces \eqn{\tau} if 
\eqn{\sigma} is the only coface of \eqn{\tau}. This function checks whether its possible to collapse \eqn{\sigma} through \eqn{\tau}, 
(if \eqn{\tau} has \eqn{\sigma} as its only coface), and if so, both simplices are removed. 
\code{tau} and \code{sigma} are sorted before comparison.
}
\references{
1. Boissonnat, Jean-Daniel, and Clement Maria. "The simplex tree: An efficient data structure for general simplicial complexes." Algorithmica 70.3 (2014): 406-427.
}
