% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{deserialize}
\alias{deserialize}
\title{Deserializes the simplex tree.}
\usage{
deserialize(complex, st = NULL)
}
\arguments{
\item{st}{a simplex tree.}
}
\description{
Provides a compressed serialization interface for the simplex tree.
}
\details{
The serialize/deserialize commands can be used to compress/uncompress the complex into 
smaller form amenable for e.g. storing on disk (see \code{saveRDS}) or saving for later use.
}
\seealso{
Other serialization: 
\code{\link{serialize}()}
}
\concept{serialization}
