% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\docType{data}
\name{empty_face}
\alias{empty_face}
\title{empty_face}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
empty_face
}
\description{
Simple alias to the NULL value, used to indicate the empty face.
}
\seealso{
traverse
}
\keyword{datasets}
