% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{faces}
\alias{faces}
\title{Face traversal}
\usage{
faces(st, sigma)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{sigma}{simplex to start the traversal at.}
}
\description{
Generates a face traversal on the simplex tree.
}
