% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{filtration}
\alias{filtration}
\title{filtration}
\usage{
filtration(st, d)
}
\arguments{
\item{st}{a simplicial complex, either as a simplex tree or an ordered list of simplices. See details.}

\item{d}{a vector of edge weights, or a 'dist' object.}
}
\description{
Creates a filtration.
}
