% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{find}
\alias{find}
\title{Find simplices}
\usage{
find(st, simplices)
}
\arguments{
\item{simplices}{a simplex or list of simplices.}
}
\value{
boolean indicating whether or not \code{simplex} exists in the tree.
}
\description{
Returns whether supplied simplices exist in the complex.
}
\details{
Traverses the simplex tree looking for \code{simplex}, returning whether or not it exists.
\code{simplex} can be specified as vector to represent a single simplex, and a list to represent a set of simplices. 
Each \code{simplex} is sorted before traversing the trie.
}
\section{Usage}{

st$find(simplex)
st$find(simplices)
}

\seealso{
insert remove
}
