% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flag}
\alias{flag}
\title{Flag complexes}
\usage{
flag(st, d)
}
\arguments{
\item{st}{a simplex tree. See details.}

\item{d}{a vector of edge weights, or a 'dist' object.}
}
\description{
Creates a filtration of flag complexes
}
\details{
A flag complex is a simplicial complex whose k-simplices for k >= 2 are completely determined 
by edges/graph of the complex. This function creates filtered simplicial complex using the supplied edge 
weights. The resulting complex is a simplex tree object endowed with additional structure; see. 
Vertices have their weights set to 0, and k-simplices w/ k >= 2 have their weights set to the maximum
weight of any of its edges.
}
\seealso{
Other simplicial complex constructors: 
\code{\link{nerve}()},
\code{\link{rips}()},
\code{\link{simplex_tree}()}
}
\concept{simplicial complex constructors}
