% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{insert_simplex}
\alias{insert_simplex}
\title{Insert simplex}
\usage{
$insert(simplex)
}
\arguments{
\item{simplex}{a k-length vector of vertex ids representing a (k-1)-simplex.}
}
\description{
Inserts a simplex.
}
\details{
This function allows insertion of arbitrary order simplices. If the simplex already exists in the tree, 
no insertion is made, and the tree is not modified. \code{simplex} is sorted before traversing the trie. 
Lower order simplices are inserted as needed if they do not already exist.
}
