% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinadic.R
\name{inverse.choose}
\alias{inverse.choose}
\title{inverse.choose}
\usage{
inverse.choose(x, k)
}
\arguments{
\item{x}{the binomial coefficient.}

\item{k}{the denominator of the binomial coefficient \code{x}.}
}
\value{
the numerator of the binomial coefficient, if the Otherwise
}
\description{
Inverts the binomial coefficient for general (n,k).
}
\details{
Given a quantity x = choose(n, k) with fixed k, finds n.
}
\examples{
100 == inverse.choose(choose(100,2), k = 2)
# TRUE 
12345 == inverse.choose(choose(12345, 5), k = 5)
# TRUE
}
