% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{k_simplices}
\alias{k_simplices}
\title{k-Simplex traversal}
\usage{
k_simplices(st, k, sigma = NULL)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{k}{the dimension of the skeleton to include.}

\item{sigma}{simplex to start the traversal at.}
}
\value{
a traversal (object of class \code{"st_traversal"}).
}
\description{
Generates a traversal on the k-simplices of the simplex tree.
}
\seealso{
Other traversals: 
\code{\link{coface_roots}()},
\code{\link{cofaces}()},
\code{\link{faces}()},
\code{\link{k_skeleton}()},
\code{\link{level_order}()},
\code{\link{link}()},
\code{\link{maximal}()},
\code{\link{preorder}()},
\code{\link{traverse}()}
}
\concept{traversals}
