% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{k_skeleton}
\alias{k_skeleton}
\title{k-Skeleton traversal}
\usage{
k_skeleton(st, k, sigma = NULL)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{k}{the dimension of the skeleton to include.}

\item{sigma}{simplex to start the traversal at.}
}
\description{
Generates a k-skeleton traversal on the simplex tree.
}
\seealso{
Other traversals: 
\code{\link{coface_roots}()},
\code{\link{cofaces}()},
\code{\link{faces}()},
\code{\link{k_simplices}()},
\code{\link{level_order}()},
\code{\link{link}()},
\code{\link{maximal}()},
\code{\link{preorder}()},
\code{\link{traverse}()}
}
\concept{traversals}
