% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{level_order}
\alias{level_order}
\title{Level order traversal}
\usage{
level_order(st, sigma = NULL)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{sigma}{simplex to start the traversal at.}
}
\description{
Generates a level order traversal on the simplex tree.
}
\seealso{
Other traversals: 
\code{\link{coface_roots}()},
\code{\link{cofaces}()},
\code{\link{faces}()},
\code{\link{k_simplices}()},
\code{\link{k_skeleton}()},
\code{\link{link}()},
\code{\link{maximal}()},
\code{\link{preorder}()},
\code{\link{traverse}()}
}
\concept{traversals}
