% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{maximal}
\alias{maximal}
\title{Maximal traversal}
\usage{
maximal(st, sigma = NULL)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{sigma}{simplex to start the traversal at.}
}
\description{
Generates a traversal on the maximal of the simplex tree.
}
