% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{plot.simplextree}
\alias{plot.simplextree}
\alias{plot.Rcpp_SimplexTree}
\title{Plots the simplex tree}
\usage{
\method{plot}{Rcpp_SimplexTree}(x, coords = NULL, vertex_opt = NULL,
  text_opt = NULL, edge_opt = NULL, polygon_opt = NULL,
  color_pal = NULL)
}
\arguments{
\item{x}{a simplex tree.}

\item{coords}{Optional (n x 2) matrix of coordinates, where n is the number of 0-simplices.}

\item{vertex_opt}{Optional parameters to modify default vertex plotting options. Passed to \code{\link[graphics]{points}}.}

\item{text_opt}{Optional parameters to modify default vertex text plotting options. Passed to \code{\link[graphics]{text}}.}

\item{edge_opt}{Optional parameters to modify default edge plotting options. Passed to \code{\link[graphics]{segments}}.}

\item{polygon_opt}{Optional parameters to modify default k-simplex plotting options for k > 1. Passed to \code{\link[graphics]{polygon}}.}

\item{color_pal}{Optional colors for each dimension of the simplicial complex.}
}
\description{
Plots the simplex tree
}
