% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{plot.Rcpp_Filtration}
\alias{plot.Rcpp_Filtration}
\title{plot.Rcpp_Filtration}
\usage{
\method{plot}{Rcpp_Filtration}(...)
}
\arguments{
\item{...}{passed to \code{\link{plot.Rcpp_SimplexTree}}}
}
\description{
plot.Rcpp_Filtration
}
\section{Functions}{
\itemize{
\item \code{plot.Rcpp_Filtration}: family of plotting methods.
}}

