% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{preorder}
\alias{preorder}
\title{Generates a preorder traversal on the simplex tree.}
\usage{
preorder(st, sigma = NULL)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{sigma}{simplex to start the traversal at.}
}
\description{
Generates a preorder traversal on the simplex tree.
}
