% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{print_cousins}
\alias{print_cousins}
\title{Prints the cousins in the simplex tree}
\usage{
print_cousins(st)
}
\description{
Prints the nodes grouped by the same last label and indexed by depth to standard out. 
By default, this is set to R's buffered output, which is shown in the R console. 
The printed format is: \cr 
\cr
(last=[label], depth=[depth of label]): [simplex] \cr
\cr
This function is useful for understanding how the simplex tree is stored, and for debugging purposes.
}
