% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{print_tree}
\alias{print_tree}
\title{Prints the simplex tree}
\usage{
print_tree(st)
}
\description{
Prints the simplicial complex to standard out. 
By default, this is set to R's buffered output, which is shown in the R console. 
The printed format is: \cr 
\cr
[vertex] (h = [subtree height]): [subtree depth]([subtree]) \cr 
\cr
Where each lists the top node (\emph{vertex}) and its corresponding subtree. The 
\emph{subtree height} displays the highest order k-simplex in that subtree. Each 
level in the subtree tree is a set of sibling k-simplices whose order is given  
by the number of dots ('.') proceeding the print level.
}
