% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{remove}
\alias{remove}
\title{Remove simplices}
\arguments{
\item{simplex}{a k-length vector of vertex ids representing a (k-1)-simplex.}
}
\description{
Removes simplices from the simplex tree. Individual simplices are specified as vectors, and a set of simplices as a list of vectors.
}
\details{
This function allows removal of a arbitrary order simplices. If \code{simplex} already exists in the tree, 
it is removed, otherwise the tree is not modified. \code{simplex} is sorted before traversing the trie.
Cofaces of \code{simplex} are also removed.
}
\section{Usage}{
 
st$remove(simplex)
}

\seealso{
find remove
}
