% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{remove_simplex}
\alias{remove_simplex}
\title{Remove simplex}
\usage{
$remove(simplex)
}
\arguments{
\item{simplex}{a k-length vector of vertex ids representing a (k-1)-simplex.}
}
\description{
Removes a simplex.
}
\details{
This function allows removal of a arbitrary order simplices. If \code{simplex} already exists in the tree, 
it is removed, otherwise the tree is not modified. \code{simplex} is sorted before traversing the trie.
Higher order simplices which are cofaces of \code{simplex} are also removed.
}
