% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-geometric.R
\name{sample-geometric}
\alias{sample-geometric}
\alias{make_geometric}
\alias{sample_unit}
\alias{sample_geometric}
\title{Sample random geometric simplicial complexes}
\usage{
make_geometric(
  d,
  radius = NULL,
  dimension = 1L,
  method = c("vietoris_rips"),
  coords = FALSE,
  ...
)

sample_unit(n, torus = FALSE, coords = FALSE)

sample_geometric(
  n,
  torus = FALSE,
  radius = NULL,
  dimension = 1L,
  method = c("vietoris_rips"),
  coords = FALSE,
  ...
)
}
\arguments{
\item{d}{a distance matrix (\code{"dist"} class) object, or a numeric matrix of
(row) coordinates of points (which will be transformed into a distance
matrix).}

\item{radius}{a numeric distance within which subsets of points will form
simplices.}

\item{dimension}{an integer maximum dimension of simplices to form.}

\item{method}{a character string indicating the model to use; matched only to
\code{"vietoris_rips"}, allowing for spaces in place of underscores,
anticipating future additional methods like \code{"cech"}.}

\item{coords}{a logical instruction to retain the coordinates from a numeric
matrix \code{d} as an attribute of the simplicial complex.}

\item{...}{additional parameters passed to the constructor indicated by
\code{method}.}

\item{n}{an integer number of starting points.}

\item{torus}{a logical instruction to identify opposite faces of the sampling
region.}
}
\value{
A \verb{[simplex_tree()]} (\verb{*_geometric()}) or a
\code{"dist"} object or coordinate matrix (\code{sample_unit()}).
}
\description{
Generate Vietoris–Rips complexes on random point clouds.
}
\details{
The geometric random graph model (see Penrose, 2003) begins with a
random sample of points from a distribution on a manifold (usually
Euclidean space), which are taken to be vertices, and introduces edges
between vertices within a fixed distance of each other.

The geometric random simplicial complex model extends this model by
constructing a Vietoris–Rips or Čech complex on the sample. See Kahle
(2011) and Bobrowski and Weinberger (2017) for key results and Kahle (2017)
for a review.
}
\examples{
set.seed(1)
## Construct geometric simplicial complexes from a sample point cloud
theta <- stats::runif(n = 24L, min = 0, max = 2*pi)
x <- cbind(x = cos(theta), y = sin(theta))
plot(x)
make_geometric(x, radius = .03, dimension = 2L)
make_geometric(x, radius = .3, dimension = 2L)
## Check distance ranges for square and toroidal samples
sqrt(2)
range(sample_unit(n = 1e3L))
sqrt(2)/2
range(sample_unit(n = 1e3L, torus = TRUE))
## Construct random geometric simplicial complexes, square and toroidal
plot(sample_geometric(24L, radius = .1, dimension = 1L))
plot(sample_geometric(24L, radius = .1, dimension = 1L, torus = TRUE))
plot(sample_geometric(24L, radius = .1, dimension = 2L))
plot(sample_geometric(24L, radius = .1, dimension = 2L, torus = TRUE))
}
\references{
Penrose M. (2003) Random Geometric Graphs. Oxford University
Press. doi:10.1093/acprof:oso/9780198506263.001.0001/acprof-9780198506263

Kahle M. (2011) Random Geometric Complexes. Discrete Comput.
Geom. 45, 553–573. doi:10.1007/s00454-010-9319-3

Bobrowski O. and Weinberger S. (2017) On the vanishing of
homology in random Čech complexes. Random Struct. Alg., 51: 14–51.
doi:10.1002/rsa.20697

Kahle M. (2017) In: J.E. Goodman, J. O'Rourke, and C.D. Tóth
(eds) Handbook of Discrete and Computational Geometry, 3rd edition. CRC
Press, Boca Raton, FL.
}
