% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{save}
\alias{save}
\alias{load}
\title{Saves/loads the simplex tree to a file}
\arguments{
\item{filename}{the filename to save/load the simplex tree to/from.}
}
\description{
Allows saving/loading the simplex tree with the help of \code{\link{readRDS}}.
}
\details{
Both saving and loading requires a filename to save the tree to. Loading
}
\examples{
st <- simplex_tree()
st$insert(1:3)
tf <- tempfile()
st$print_tree()
st$save(tf)
st$clear()
print(st)
< empty simplex tree >
st$load(tf)
st$print_tree()
# 1 (h = 2): .( 2 3 )..( 3 )
# 2 (h = 1): .( 3 )
# 3 (h = 0):
}
\seealso{
serialize deserialize
}
