% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{serialize}
\alias{serialize}
\alias{deserialize}
\title{Serializes/Deserializes the simplex tree.}
\arguments{
\item{x}{a list of simplices to insert into the tree.}
}
\description{
Provides basic deserialization interface for the simplex tree.
}
\details{
Saves the simplex tree as a compressed RDS file with \code{\link{saveRDS}}. Only the (generally higher order) 
simplices which have themselves as a unique coface are saved.
}
\examples{
st <- simplex_tree()
st$insert(c(1, 2, 3))
tmp <- st$serialize()
print(tmp)
# [[1]]
# [1] 1 2 3
st$clear()
st$deserialize(tmp)
st$print_tree()
}
