% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplextree-package.R
\docType{package}
\name{simplextree-package}
\alias{simplextree-package}
\title{simplextree package}
\description{
Provides an R/Rcpp implementation of a Simplex Tree dsat structure and its related tools.
}
\author{
Matt Piekenbrock
}
