% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{straverse}
\alias{straverse}
\title{straverse}
\usage{
straverse(traversal, f, ...)
}
\description{
straverse
}
\seealso{
Other traversals: 
\code{\link{ltraverse}()},
\code{\link{traverse}()}
}
\concept{traversals}
