% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{threshold}
\alias{threshold}
\title{Filtered complex thresholding}
\usage{
threshold(st, index = NULL, value = NULL)
}
\arguments{
\item{st}{simplex tree.}

\item{index}{integer index to threshold to.}

\item{value}{numeric index to threshold filtration.}
}
\description{
Thresholds a given filtered simplicial complex.
}
