% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{straverse}
\alias{straverse}
\title{straverse}
\usage{
straverse(traversal, f, ...)
}
\arguments{
\item{traversal}{the type of traversal.}

\item{f}{the function to apply to each simplex.}

\item{...}{unused.}
}
\description{
straverse
}
\section{Functions}{
\itemize{
\item \code{straverse}: \code{\link{sapply}} version of \code{\link{traverse}}.
}}

\seealso{
Other traversals: 
\code{\link{ltraverse}()},
\code{\link{traverse}()}
}
\concept{traversals}
