% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-union-find.R
\docType{class}
\name{union_find}
\alias{union_find}
\title{Union-find}
\usage{
union_find(n = 0L)
}
\arguments{
\item{n}{Number of elements in the set.}
}
\value{
A disjoint set, as a \code{Rcpp_UnionFind} object (Rcpp module).
}
\description{
Union find structure exposed as an Rcpp Module.
}
\section{Methods}{

\describe{
\item{\verb{$print.simplextree}}{
S3 method to print a basic summary of the simplex tree.
}
}
}

\author{
Matt Piekenbrock
}
