% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{adjacent}
\alias{adjacent}
\title{Adjacent vertices.}
\description{
Returns a vector of vertex ids that are immediately adjacent to a given vertex.
}
\examples{
st <- simplex_tree()
st$insert(1:3)
st$adjacent(2) ## 1 3
}
