% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices.R
\name{adjacent}
\alias{adjacent}
\title{Adjacent vertices}
\usage{
adjacent(st, vertices)
}
\arguments{
\item{st}{a simplex tree.}

\item{vertices}{vertex ids.}
}
\value{
a list of double vectors of vertices adjacent to each of \code{vertices} in \code{st} (or \code{numeric(0)} for vertices not in \code{st}), unlisted to a single vector if \verb{length(vertices == 1}.
}
\description{
Returns a vector of vertex ids that are immediately adjacent to a given vertex.
}
\examples{
st <- simplex_tree(1:3)
st \%>\% adjacent(2) 
# 1 3
}
\seealso{
Other vertex-level operations: 
\code{\link{degree}()},
\code{\link{reindex}()}
}
\concept{vertex-level operations}
