% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{clear}
\alias{clear}
\title{Clears the simplex tree}
\usage{
clear(st)
}
\description{
Removes all simplices from the simplex tree, except the root node.
}
\examples{
st <- simplex_tree()
st$insert(1:3)
print(st) ## Simplex Tree with (3, 3, 1) (0, 1, 2)-simplices
st$clear()
print(st) ## < empty simplex tree >
}
