% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{clear}
\alias{clear}
\title{Clear a simplex tree}
\usage{
clear(st)
}
\arguments{
\item{st}{a simplex tree object.}
}
\value{
the simplex tree \code{st} with simplices removed, invisibly.
}
\description{
Removes all simplices from the simplex tree, except the root
node.
}
\examples{
st <- simplex_tree()
st \%>\% insert(1:3)
print(st) ## Simplex Tree with (3, 3, 1) (0, 1, 2)-simplices
st \%>\% clear()
print(st) ## < empty simplex tree >
}
\seealso{
Other complex-level operations: 
\code{\link{contract}()},
\code{\link{expand}()},
\code{\link{threshold}()}
}
\concept{complex-level operations}
