% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{cofaces}
\alias{cofaces}
\title{Coface traversal}
\usage{
cofaces(st, sigma)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{sigma}{simplex to start the traversal at.}
}
\description{
Generates a coface traversal on the simplex tree.
}
