% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{degree}
\alias{degree}
\title{Vertex degree}
\usage{
degree(st, vertices)
}
\arguments{
\item{st}{a simplex tree.}

\item{vertices}{the vertex ids to check the degree of.}
}
\description{
Returns the number of edges (degree) for each given vertex id.
}
