% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{deserialize}
\alias{deserialize}
\title{Deserialize the simplex tree}
\usage{
deserialize(complex, st = NULL)
}
\arguments{
\item{complex}{The result of \code{\link{serialize}}.}

\item{st}{optionally, the simplex tree to insert into. Otherwise a new one is created.}
}
\description{
Provides a compressed serialization interface for the simplex tree.
}
\details{
The serialize/deserialize commands can be used to compress/uncompress the complex into 
smaller form amenable for e.g. storing on disk (see \code{saveRDS}) or saving for later use.
}
\seealso{
Other serialization: 
\code{\link{serialize}()}
}
\concept{serialization}
