% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\docType{data}
\name{empty_face}
\alias{empty_face}
\title{Empty faces}
\format{
An object of class \code{integer} of length 0.
}
\usage{
empty_face
}
\description{
Alias to the empty integer vector (integer(0L)). Used to indicate the empty face of the tree.
}
\seealso{
traverse
}
\keyword{datasets}
