% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rips.R
\name{enclosing_radius}
\alias{enclosing_radius}
\title{enclosing_radius}
\usage{
enclosing_radius(d)
}
\arguments{
\item{d}{a \code{\link[stats]{dist}} object.}
}
\description{
Computes the enclosing radius of a set of distances.
}
\details{
The enclosing radius is useful as an upper bound of the scale parameter 
for the rips filtration. Scales above the enclosing radius render the Vietoris–Rips
complex as a simplicial cone, beyond which the homology is trivial.
}
