% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{find}
\alias{find}
\title{Find simplices}
\usage{
$find(simplex)
}
\arguments{
\item{simplex}{a k-length vector of vertex ids representing a (k-1)-simplex.}
}
\value{
boolean indicating whether or not \code{simplex} exists in the tree.
}
\description{
Finds whether simplices exist the simplex tree. Use a vector to represent a simplex, and a list to represent a set of simplices.
}
\details{
Traverses the simplex tree looking for \code{simplex}, returning whether or not it exists. 
\code{simplex} is sorted before traversing the trie.
}
