% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{find}
\alias{find}
\title{Find simplices}
\usage{
find(simplex), find(simplices)
}
\arguments{
\item{simplex}{a k-length vector of vertex ids representing a (k-1)-simplex. Individual simplices are specified as vectors, and a set of simplices as a list of vectors.}

\item{simplices}{a list of simplices.}
}
\value{
boolean indicating whether or not \code{simplex} exists in the tree.
}
\description{
Finds whether simplices exist the simplex tree.
}
\details{
Traverses the simplex tree looking for \code{simplex}, returning whether or not it exists.
\code{simplex} can be specified as vector to represent a single simplex, and a list to represent a set of simplices. 
\code{simplex} is sorted before traversing the trie.
}
\seealso{
find remove
}
