% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{insert}
\alias{insert}
\title{Insert simplices}
\usage{
insert(simplex), insert(simplices)
}
\arguments{
\item{simplex}{a k-length vector of vertex ids representing a (k-1)-simplex.}

\item{simplices}{a list of simplices.}
}
\description{
Inserts simplices into the simplex tree. Individual simplices are specified as vectors, and a set of simplices as a list of vectors.
}
\details{
This function allows insertion of arbitrary order simplices. If the simplex already exists in the tree, 
no insertion is made, and the tree is not modified. \code{simplex} is sorted before traversing the trie. 
Faces of \code{simplex} not in the simplex tree are inserted as needed.
}
\examples{
st <- simplex_tree()
st$insert(1:3) ## inserts the 2-simplex { 1, 2, 3 }
st$insert(list(4:5, 6)) ## inserts a 1-simplex { 4, 5 } and a 0-simplex { 6 }.
}
\seealso{
find remove
}
