% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{is_face}
\alias{is_face}
\title{Face test}
\usage{
is_face(st, tau, sigma)
}
\arguments{
\item{st}{a simplex tree.}

\item{tau}{a simplex which may contain \code{sigma} as a coface.}

\item{sigma}{a simplex which may contain \code{tau} as a face.}
}
\value{
boolean indicating whether \code{tau} is a face of \code{sigma}.
}
\description{
Checks whether a simplex is a face of another simplex and is in the complex.
}
\details{
A simplex \eqn{\tau} is a face of \eqn{\sigma} if the vertices of \eqn{\tau} are vertices of \eqn{\sigma}. This function
checks whether that is true. \code{tau} and \code{sigma} are sorted before comparison.
}
\examples{
st <- simplex_tree()
st \%>\% insert(1:3)
st \%>\% is_face(2:3, 1:3)
st \%>\% is_face(1:3, 2:3)
}
\seealso{
\href{https://en.cppreference.com/w/cpp/algorithm/includes}{std::includes}
}
