% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_tree}
\alias{is_tree}
\title{Tree (acyclicity) test}
\usage{
is_tree(st)
}
\arguments{
\item{st}{a simplex tree.}
}
\value{
a boolean indicating whether \code{st} is acyclic.
}
\description{
This function performs a breadth-first search on the simplicial complex, checking if the complex is acyclic.
}
\examples{
st <- simplex_tree()
st \%>\% insert(list(1:2, 2:3))
st \%>\% is_tree() # true
st \%>\% insert(c(1, 3))
st \%>\% is_tree() # false
}
