% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{k_simplices}
\alias{k_simplices}
\title{Generates a traversal on the k-simplices of the simplex tree.}
\usage{
k_simplices(st, k, sigma = NULL)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{k}{the dimension of the skeleton to include.}

\item{sigma}{simplex to start the traversal at.}
}
\description{
Generates a traversal on the k-simplices of the simplex tree.
}
