% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{k_skeleton}
\alias{k_skeleton}
\title{Generates a k-skeleton traversal on the simplex tree.}
\usage{
k_skeleton(st, k, sigma = NULL)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{k}{the dimension of the skeleton to include.}

\item{sigma}{simplex to start the traversal at.}
}
\description{
Generates a k-skeleton traversal on the simplex tree.
}
