% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{level_order}
\alias{level_order}
\title{Level order traversal}
\usage{
level_order(st, sigma = NULL)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{sigma}{simplex to start the traversal at.}
}
\description{
Generates a level order traversal on the simplex tree.
}
