% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{link}
\alias{link}
\title{Link traversal}
\usage{
link(st, sigma)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{sigma}{simplex to start the traversal at.}
}
\value{
a traversal (object of class \code{"st_traversal"}).
}
\description{
Generates a traversal on the link of a given simplex in the
simplex tree.
}
\seealso{
Other traversals: 
\code{\link{coface_roots}()},
\code{\link{cofaces}()},
\code{\link{faces}()},
\code{\link{k_simplices}()},
\code{\link{k_skeleton}()},
\code{\link{level_order}()},
\code{\link{maximal}()},
\code{\link{preorder}()},
\code{\link{traverse}()}
}
\concept{traversals}
