% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{link}
\alias{link}
\title{Generates a traversal on the link of a given simplex in the simplex tree.}
\usage{
link(st, sigma)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{sigma}{simplex to start the traversal at.}
}
\description{
Generates a traversal on the link of a given simplex in the simplex tree.
}
