% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{ltraverse}
\alias{ltraverse}
\title{ltraverse}
\usage{
ltraverse(traversal, f, ...)
}
\arguments{
\item{traversal}{the type of traversal.}

\item{f}{the function to apply to each simplex.}

\item{...}{unused.}
}
\description{
ltraverse
}
\seealso{
Other traversals: 
\code{\link{straverse}()},
\code{\link{traverse}()}
}
\concept{traversals}
