% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nerve.R
\name{nerve}
\alias{nerve}
\title{nerve}
\usage{
nerve(st, cover, k = st$dimension, threshold = 1L, neighborhood = NULL)
}
\arguments{
\item{st}{a simplex tree.}

\item{cover}{list of integers indicating set membership. See details.}

\item{k}{max simplex dimension to consider.}

\item{threshold}{the number of elements in common for \code{k} sets to be considered intersecting. Defaults to 1.}

\item{neighborhood}{which combinations of sets to check. See details.}
}
\description{
Compute the nerve of a given cover.
}
\details{
This computes the nerve of a given cover, adding a \emph{k}-simplex for each combination of \emph{k+1} sets 
in the given \code{cover} that have at least \code{threshold} elements in their common intersection.
}
