% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{preorder}
\alias{preorder}
\title{Preorder traversal}
\usage{
preorder(st, sigma = NULL)
}
\arguments{
\item{st}{the simplex tree to traverse.}

\item{sigma}{simplex to start the traversal at.}
}
\description{
Generate a preorder traversal on the simplex tree.
}
\seealso{
Other traversals: 
\code{\link{coface_roots}()},
\code{\link{cofaces}()},
\code{\link{faces}()},
\code{\link{k_simplices}()},
\code{\link{k_skeleton}()},
\code{\link{level_order}()},
\code{\link{link}()},
\code{\link{maximal}()},
\code{\link{traverse}()}
}
\concept{traversals}
