% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{print_simplices}
\alias{print_simplices}
\title{Prints simplices in a formatted way}
\usage{
print_simplices(x, format = c("short", "column", "row"))
}
\description{
Prints a traversal, a simplex tree, or a list of simplices to the R console, with 
options to customize how the simplices are printed.
}
