% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{remove}
\alias{remove}
\title{Remove simplices}
\usage{
$remove(simplex)
}
\arguments{
\item{simplex}{a k-length vector of vertex ids representing a (k-1)-simplex.}
}
\description{
Removes simplices from the simplex tree. Use a vector to represent a simplex, and a list to represent a set of simplices.
}
\details{
This function allows removal of a arbitrary order simplices. If \code{simplex} already exists in the tree, 
it is removed, otherwise the tree is not modified. \code{simplex} is sorted before traversing the trie.
Cofaces of \code{simplex} are also removed.
}
