% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rips.R
\name{rips}
\alias{rips}
\title{Vietoris–Rips complex}
\usage{
rips(d, eps = enclosing_radius(d), dim = 1L, filtered = FALSE)
}
\arguments{
\item{d}{a numeric 'dist' vector.}

\item{eps}{diameter parameter.}

\item{dim}{maximum dimension to construct up to. Defaults to 1 (edges only).}

\item{filtered}{whether to construct the filtration. Defaults to false. See details.}
}
\description{
Constructs the Vietoris–Rips complex.
}
\seealso{
Other simplicial complex constructors: 
\code{\link{flag}()},
\code{\link{nerve}()},
\code{\link{simplex_tree}()}
}
\concept{simplicial complex constructors}
