% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{serialize}
\alias{serialize}
\title{Serializes/deserializes the simplex tree.}
\usage{
serialize(st)
}
\arguments{
\item{st}{a simplex tree.}
}
\description{
Provides a compressed serialization interface for the simplex tree.
}
\details{
The serialize/deserialize commands can be used to compress/uncompress the complex into 
smaller form amenable for e.g. storing on disk (see \code{saveRDS}) or saving for later use. 
The serialization
}
\examples{
st <- simplex_tree(list(1:5, 7:9))
tmp <- serialize(st)
print(tmp)
# [[1]]
# [1] 1 2 3
st$clear()
st$deserialize(tmp)
st$print_tree()
}
\concept{serialization}
