% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{serialize}
\alias{serialize}
\title{Serializes/deserializes the simplex tree.}
\usage{
serialize(st)
}
\arguments{
\item{st}{a simplex tree.}
}
\description{
Provides basic serialization interface for the simplex tree.
}
\details{
Saves the simplex tree as a compressed RDS file with \code{\link{saveRDS}}. Only the (generally higher order) 
simplices which have themselves as a unique coface are saved.
}
\examples{
st <- simplex_tree()
st$insert(c(1, 2, 3))
tmp <- st$serialize()
print(tmp)
# [[1]]
# [1] 1 2 3
st$clear()
st$deserialize(tmp)
st$print_tree()
}
\concept{serialization}
