% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplextree-package.R
\docType{package}
\name{simplextree-package}
\alias{simplextree-package}
\alias{simplextree}
\title{'simplextree' package}
\description{
Provides an R/Rcpp implementation of a Simplex Tree data structure and its related tools.
}
\details{
This package provides a lightweight implementation of a Simplex Tree data structure, exported as an Rcpp Module.
The current implementation provides a limited API and a subset of the functionality described in the paper.
}
\author{
Matt Piekenbrock
}
