% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_tree.R
\name{traversals}
\alias{traversals}
\alias{print.st_traversal}
\alias{as.list.st_traversal}
\title{Methods for traversal objects}
\usage{
\method{print}{st_traversal}(x, ...)

\method{as.list}{st_traversal}(x, ...)
}
\arguments{
\item{x}{traversal object.}

\item{...}{unused.}
}
\description{
Methods for traversal objects
}
