% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_find.R
\docType{class}
\name{union_find}
\alias{union_find}
\title{UnionFind}
\usage{
union_find(n = 0L)
}
\arguments{
\item{n}{Number of elements in the set.}
}
\value{
A disjoint set, as a \code{Rcpp_UnionFind} object (Rcpp module).
}
\description{
Union find structure exposed as an Rcpp Module.
}
\section{Methods}{
 
\describe{
    \item{$\code{print.simplextree}}{ S3 method to print a basic summary of the simplex tree. }
}
}

\author{
Matt Piekenbrock
}
